Week 4/aie3-midterm/requirements.txt
chainlit==0.7.700
langchain==0.2.5
langchain_community==0.2.5
langchain_core==0.2.9
langchain-openai==0.1.9
langchain_text_splitters==0.2.1
python-dotenv==1.0.1
pymupdf==1.24.5
python-socketio==5.11.2
openai==1.35.0
qdrant-client==1.9.2
